+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [6.5.1](https://github.com/mll-lab/php-utils/compare/v6.5.0...v6.5.1) (2026-03-11)
+
+### Bug Fixes
+
+* extract flowcell ID, handle backslash paths, support 6-digit dates ([#81](https://github.com/mll-lab/php-utils/issues/81)) ([830068d](https://github.com/mll-lab/php-utils/commit/830068dc6d3b39616677eef2ce21407a2801b969))
+
+## [6.5.0](https://github.com/mll-lab/php-utils/compare/v6.4.0...v6.5.0) (2026-03-11)
+
+### Features
+
+* Implement IlluminaRunFolder - We need to Parse the run folder to get sequencing run specific flowcell to identify run ([aed2772](https://github.com/mll-lab/php-utils/commit/aed27725848d09cd569d36e88e79bcb3a6ff2f05))
+
 ## [6.4.0](https://github.com/mll-lab/php-utils/compare/v6.3.0...v6.4.0) (2026-03-09)
 
 ### Features